% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{ac_accounts}
\alias{ac_accounts}
\title{ActiveCampaign Accounts}
\usage{
ac_accounts(search = NULL, .progress = NULL)
}
\arguments{
\item{search}{Search string}

\item{.progress}{Optional progressr callback}
}
\value{
A tibble of accounts
}
\description{
Operations for ActiveCampaign accounts (companies).
}
