% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-tags.R
\name{ac_add_tag}
\alias{ac_add_tag}
\title{Add a Tag to a Contact}
\usage{
ac_add_tag(contact_id, tag_id)
}
\arguments{
\item{contact_id}{Contact ID}

\item{tag_id}{Tag ID}
}
\value{
A single-row tibble
}
\description{
Add a Tag to a Contact
}
