% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ac_cache_file}
\alias{ac_cache_file}
\title{Get the Cache File Path for a Key}
\usage{
ac_cache_file(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
File path
}
\description{
Get the Cache File Path for a Key
}
\keyword{internal}
