% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ac_cache_path}
\alias{ac_cache_path}
\title{Get or Set the Cache Directory}
\usage{
ac_cache_path(path = NULL)
}
\arguments{
\item{path}{New cache directory path (optional). If \code{NULL}, returns
the current path.}
}
\value{
The cache directory path (invisibly if setting)
}
\description{
Get or Set the Cache Directory
}
