% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{ac_campaign_messages}
\alias{ac_campaign_messages}
\title{Get Campaign Messages}
\usage{
ac_campaign_messages(campaign_id)
}
\arguments{
\item{campaign_id}{Campaign ID}
}
\value{
A tibble of messages for the campaign
}
\description{
Get Campaign Messages
}
