% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{ac_campaigns}
\alias{ac_campaigns}
\title{ActiveCampaign Campaigns}
\usage{
ac_campaigns(.progress = NULL)
}
\arguments{
\item{.progress}{Optional progressr callback}
}
\value{
A tibble of campaigns
}
\description{
Fetch campaign data and performance metrics.
}
