% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ac_check_auth}
\alias{ac_check_auth}
\title{Check Authentication Status}
\usage{
ac_check_auth()
}
\value{
Invisibly returns \code{TRUE} if authenticated
}
\description{
Check Authentication Status
}
\keyword{internal}
