% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ac_coerce_types}
\alias{ac_coerce_types}
\title{Coerce Common Column Types}
\usage{
ac_coerce_types(df)
}
\arguments{
\item{df}{A tibble}
}
\value{
A tibble with coerced types
}
\description{
Converts date columns to POSIXct and keeps IDs as character.
}
\keyword{internal}
