% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-custom-fields.R
\name{ac_contact_custom_fields}
\alias{ac_contact_custom_fields}
\title{ActiveCampaign Contact Custom Fields}
\usage{
ac_contact_custom_fields()
}
\value{
A tibble of field definitions
}
\description{
Fetch and pivot custom field data for contacts.
}
