% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-custom-fields.R
\name{ac_contact_custom_fields_wide}
\alias{ac_contact_custom_fields_wide}
\title{Get Contact Custom Fields in Wide Format}
\usage{
ac_contact_custom_fields_wide(contact_id = NULL)
}
\arguments{
\item{contact_id}{Optional contact ID(s) to filter by}
}
\value{
A tibble with \code{contact_id} + one column per custom field
}
\description{
Get Contact Custom Fields in Wide Format
}
