% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-tags.R
\name{ac_contact_tags}
\alias{ac_contact_tags}
\title{ActiveCampaign Contact Tags}
\usage{
ac_contact_tags(contact_id)
}
\arguments{
\item{contact_id}{Contact ID}
}
\value{
A tibble of tags
}
\description{
Manage tags on contacts.
}
