% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac_contact_url}
\alias{ac_contact_url}
\title{Build an ActiveCampaign Contact URL}
\usage{
ac_contact_url(contact_id)
}
\arguments{
\item{contact_id}{Contact ID(s)}
}
\value{
Character vector of contact URLs
}
\description{
Build an ActiveCampaign Contact URL
}
\keyword{internal}
