% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{ac_create_account}
\alias{ac_create_account}
\title{Create an Account}
\usage{
ac_create_account(name, url = NULL, ...)
}
\arguments{
\item{name}{Account name (required)}

\item{url}{Account website URL}

\item{...}{Additional fields}
}
\value{
A single-row tibble
}
\description{
Create an Account
}
