% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{ac_create_contact}
\alias{ac_create_contact}
\title{Create a Contact}
\usage{
ac_create_contact(
  email,
  first_name = NULL,
  last_name = NULL,
  phone = NULL,
  ...
)
}
\arguments{
\item{email}{Email address (required)}

\item{first_name}{First name}

\item{last_name}{Last name}

\item{phone}{Phone number}

\item{...}{Additional fields}
}
\value{
A single-row tibble of the created contact
}
\description{
Create a Contact
}
