% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{ac_create_tag}
\alias{ac_create_tag}
\title{Create a Tag}
\usage{
ac_create_tag(name, type = "contact", description = NULL)
}
\arguments{
\item{name}{Tag name}

\item{type}{Tag type: \code{"contact"} or \code{"deal"}}

\item{description}{Tag description}
}
\value{
A single-row tibble
}
\description{
Create a Tag
}
