% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{ac_create_task}
\alias{ac_create_task}
\title{Create a Task}
\usage{
ac_create_task(
  title,
  deal_id = NULL,
  due_date = NULL,
  type = NULL,
  assignee = NULL,
  note = NULL,
  ...
)
}
\arguments{
\item{title}{Task title}

\item{deal_id}{Deal ID to associate with}

\item{due_date}{Due date (character \code{"YYYY-MM-DD"} or Date)}

\item{type}{Task type ID}

\item{assignee}{User ID to assign to}

\item{note}{Task note/body}

\item{...}{Additional fields}
}
\value{
A single-row tibble
}
\description{
Create a Task
}
