% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{ac_create_webhook}
\alias{ac_create_webhook}
\title{Create a Webhook}
\usage{
ac_create_webhook(name, url, events, sources = "0")
}
\arguments{
\item{name}{Webhook name}

\item{url}{URL to receive webhook POSTs}

\item{events}{Character vector of event names (e.g., \code{"deal_add"},
\code{"contact_add"}, \code{"deal_update"})}

\item{sources}{Character vector of source IDs (e.g., \code{"0"} for all)}
}
\value{
A single-row tibble
}
\description{
Create a Webhook
}
