% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal-pipelines.R
\name{ac_deal_pipelines}
\alias{ac_deal_pipelines}
\title{ActiveCampaign Deal Pipelines and Stages}
\usage{
ac_deal_pipelines()
}
\value{
A tibble of pipeline definitions
}
\description{
Fetch pipeline and stage definitions.
}
