% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal-pipelines.R
\name{ac_deal_stages}
\alias{ac_deal_stages}
\title{List Deal Stages}
\usage{
ac_deal_stages(pipeline = NULL)
}
\arguments{
\item{pipeline}{Filter by pipeline (group) ID (optional)}
}
\value{
A tibble of stage definitions
}
\description{
List Deal Stages
}
