% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac_deal_url}
\alias{ac_deal_url}
\title{Build an ActiveCampaign Deal URL}
\usage{
ac_deal_url(deal_id)
}
\arguments{
\item{deal_id}{Deal ID(s)}
}
\value{
Character vector of deal URLs
}
\description{
Build an ActiveCampaign Deal URL
}
\keyword{internal}
