% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ac_delete_one}
\alias{ac_delete_one}
\title{Perform a DELETE Request}
\usage{
ac_delete_one(endpoint)
}
\arguments{
\item{endpoint}{API endpoint (e.g., \code{"deals/123"})}
}
\value{
Invisibly returns \code{TRUE} on success
}
\description{
Perform a DELETE Request
}
\keyword{internal}
