% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{ac_delete_webhook}
\alias{ac_delete_webhook}
\title{Delete a Webhook}
\usage{
ac_delete_webhook(id)
}
\arguments{
\item{id}{Webhook ID}
}
\value{
Invisibly returns \code{TRUE}
}
\description{
Delete a Webhook
}
