% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ac_get_one}
\alias{ac_get_one}
\title{Perform a Single-Entity GET Request}
\usage{
ac_get_one(endpoint, entity_key)
}
\arguments{
\item{endpoint}{API endpoint (e.g., \code{"deals/123"})}

\item{entity_key}{Response key (e.g., \code{"deal"})}
}
\value{
A single-row tibble
}
\description{
Perform a Single-Entity GET Request
}
\keyword{internal}
