% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ac_post_one}
\alias{ac_post_one}
\title{Perform a Create (POST) Request}
\usage{
ac_post_one(endpoint, entity_key, body)
}
\arguments{
\item{endpoint}{API endpoint}

\item{entity_key}{Wrapper key for the body (e.g., \code{"deal"})}

\item{body}{Named list of fields}
}
\value{
A single-row tibble of the created entity
}
\description{
Perform a Create (POST) Request
}
\keyword{internal}
