% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ac_put_one}
\alias{ac_put_one}
\title{Perform an Update (PUT) Request}
\usage{
ac_put_one(endpoint, entity_key, body)
}
\arguments{
\item{endpoint}{API endpoint (e.g., \code{"deals/123"})}

\item{entity_key}{Wrapper key for the body}

\item{body}{Named list of fields to update}
}
\value{
A single-row tibble of the updated entity
}
\description{
Perform an Update (PUT) Request
}
\keyword{internal}
