% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-tags.R
\name{ac_remove_tag}
\alias{ac_remove_tag}
\title{Remove a Tag from a Contact}
\usage{
ac_remove_tag(contact_tag_id)
}
\arguments{
\item{contact_tag_id}{The contact-tag association ID (from \code{\link[=ac_contact_tags]{ac_contact_tags()}})}
}
\value{
Invisibly returns \code{TRUE}
}
\description{
Remove a Tag from a Contact
}
