% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{ac_tasks}
\alias{ac_tasks}
\title{ActiveCampaign Tasks}
\usage{
ac_tasks(.progress = NULL)
}
\arguments{
\item{.progress}{Optional progressr callback}
}
\value{
A tibble of tasks
}
\description{
CRUD operations for ActiveCampaign tasks.
}
