% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{ac_trigger_automation}
\alias{ac_trigger_automation}
\title{Trigger an Automation for a Contact}
\usage{
ac_trigger_automation(automation_id, contact_id)
}
\arguments{
\item{automation_id}{Automation ID}

\item{contact_id}{Contact ID}
}
\value{
A single-row tibble of the contact-automation association
}
\description{
Adds a contact to an automation.
}
