% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{ac_update_contact}
\alias{ac_update_contact}
\title{Update a Contact}
\usage{
ac_update_contact(id, ...)
}
\arguments{
\item{id}{Contact ID}

\item{...}{Fields to update}
}
\value{
A single-row tibble of the updated contact
}
\description{
Update a Contact
}
